refactor: changed from BackupHolder to SyncHolder
