+++ app/src/main/java/eu/kanade/tachiyomi/data/sync/SyncHolder.kt
@@ -1,7 +1,7 @@
-package eu.kanade.tachiyomi.data.backup
+package eu.kanade.tachiyomi.data.sync
 
 import eu.kanade.tachiyomi.data.backup.models.Backup
 
-object BackupHolder {
+object SyncHolder {
     var backup: Backup? = null
 }
